tests/wpt/tests/html/semantics/forms/the-input-element/password-delete-space.html

<!DOCTYPE html>
<title>Backspace with trailing white space in password field</title>
<link rel="help" href="https://bugs.chromium.org/p/chromium/issues/detail?id=1400844">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#password-state-%28type=password%29">
<link rel="author" href="mailto:xiaochengh@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<input id="target" type="password" value="   ">

<script>
promise_test(async () => {
  target.focus();
  target.selectionStart = 2;
  await test_driver.send_keys(target, '\uE003');
  assert_equals(target.value, "  ");
}, "Backspace with trailing white space in password field");
</script>
