LayoutTests/platform/ios/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/password-delete-space-expected.txt



FAIL Backspace with trailing white space in password field assert_equals: expected "  " but got "   "
